2 rest controllers methods working

+++ src/main/java/com/spring/freezer/repositories/FoodRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.spring.freezer.entities.Food;
 
-public interface FoodRepository extends JpaRepository<Food, Long> {
+public interface FoodRepository extends JpaRepository<Food, Integer> {
 }
